Make library names unique
Previously, the built library names were identical to the Velodyne driver
library names, and they were also installed in the global lib directory.
This means they would cause a conflict with the Velodyne binary packages
and it would be impossible to install both it and this package at the
same time.  This makes the library names unique, but the nodelet names
are still the same, so it should be backwards-compatible.